feat: track end position of genetic map regions
certain genetic map formats (e.g. bedgraph, bigwig) specify an end position of their rate estimation
range. this information is only required at the end of chromosomes, in order to interpolate the
final genetic position value that is then extended to all positions past the end of the estimation
range.