feat: change bed region query output back to bolt format
bedgraph output doesn't allow the type of rate spikes that I'd like to be able to introduce in
certain experimental contexts. as mentioned in a prior commit, using bolt style output for this type
of query requires additional, fake annotations to set the end chromosome region rates and genetic
positions to their appropriate values; the tracking required to allow this functionality is now
added. bed query tests are updated to reflect this new behavior.